backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Imagegrab==0.0.3
jiter==0.6.1
openai==1.52.0
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.32.0
